requirements.txt
python-telegram-bot
python-dotenv
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client

